requirements.txt
arena-api @ file:///C:/path_to_file/arena_api-2.7.1-py3-none-any/arena_api-2.7.1-py3-none-any.whl
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.57.0
fsspec==2024.6.1
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.2.1
Nuitka==2.7
numpy==2.0.2
opencv-contrib-python==4.11.0.86
ordered-set==4.1.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
PySDL2==0.9.17
pysdl2-dll==2.32.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2024.8.30
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
zipp==3.21.0
zstandard==0.23.0
